% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_team_abbrs}
\alias{clean_team_abbrs}
\title{Standardize NBA Team Abbreviations}
\usage{
clean_team_abbrs(abbr, keep_non_matches = TRUE)
}
\arguments{
\item{abbr}{a character vector of abbreviations}

\item{keep_non_matches}{If \code{TRUE} (the default) an element of \code{abbr} that can't
be matched to any of the internal mapping vectors will be kept as is. Otherwise
it will be replaced with \code{NA}.}
}
\value{
A character vector with the length of \code{abbr} and cleaned team abbreviations
if they are included in \code{\link{team_abbr_mapping}}. Non matches may be replaced
with \code{NA} (depending on the value of \code{keep_non_matches}).
}
\description{
This function standardizes NBA team abbreviations to ESPN defaults.
}
\examples{
x <- c("ALT", "BKN", "BRK", "BROK", "UTAH", "UTA", "UTAA")

# keep non matches
nbaplotR::clean_team_abbrs(x)

# replace non matches
nbaplotR::clean_team_abbrs(x, keep_non_matches = FALSE)
}
