% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nba_logos.R, R/nbaplotR-package.R
\docType{data}
\name{GeomNBAlogo}
\alias{GeomNBAlogo}
\alias{nbaplotR}
\alias{nbaplotR-package}
\title{nbaplotR: What the Package Does (One Line, Title Case)}
\format{
An object of class \code{GeomNFLlogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomNBAlogo
}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{datasets}
\keyword{internal}
