% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid NBA or WNBA Team Abbreviations}
\usage{
valid_team_names(league = c("NBA", "WNBA"))
}
\arguments{
\item{league}{One of \code{"NBA"} or \code{"WNBA"}}
}
\value{
A vector of type \code{"character"}.
}
\description{
The abbreviations used in this function are extracted from ESPN
}
\examples{
# List valid NBA team abbreviations
valid_team_names("NBA")

# List valid WNBA team abbreviations
valid_team_names("WNBA")
}
