% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wnba_team_abbr_mapping}
\alias{wnba_team_abbr_mapping}
\title{Alternate WNBA Team Abbreviation Mapping}
\format{
A named character vector
\describe{
\item{name attribute}{The "alternate" name.}
\item{value attribute}{The "correct" name.}
}
}
\usage{
wnba_team_abbr_mapping
}
\description{
A named character vector mapping common alternate team abbreviations.
}
\details{
You can suggest additions to this table by
\href{https://github.com/mrcaseb/nbaplotR/issues/new/choose}{opening an issue in nbaplotR}.
}
\examples{
\donttest{
wnba_team_abbr_mapping[c("CONN", "DAL", "SEA")]
}

}
\keyword{datasets}
