% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nba_logos.R, R/nbaplotR-package.R
\docType{data}
\name{GeomNBAlogo}
\alias{GeomNBAlogo}
\alias{nbaplotR-package}
\title{nbaplotR: NBA Logo Plots in 'ggplot2'}
\format{
An object of class \code{GeomNBAlogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomNBAlogo
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions to visualize National Basketball League analysis in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrcaseb/nbaplotR}
  \item \url{https://mrcaseb.github.io/nbaplotR/}
  \item Report bugs at \url{https://github.com/mrcaseb/nbaplotR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{datasets}
\keyword{internal}
