% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid NAB Team Abbreviations}
\usage{
valid_team_names()
}
\value{
A vector of type \code{"character"}.
}
\description{
The abbreviations used in this function are extracted from ESPN
}
\examples{
# List valid team abbreviations
valid_team_names()
}
